# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple todo rest api",
        "title": "Todo REST API",
        "contact": {
            "email": "ibuki420v@gmail.com"
        },
        "license": {
            "name": "MIT LICENSE",
            "url": "https://github.com/v420v/TodoApp/blob/main/LICENSE.md"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/todos": {
            "get": {
                "description": "get a list of TODOs that are not deleted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Get Todo list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todos.Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "insert todo to db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Post todo",
                "parameters": [
                    {
                        "description": "Todo object",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todos.Todo"
                        }
                    }
                }
            }
        },
        "/todos/{id}/delete": {
            "post": {
                "description": "delete todo by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Delete todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "todos.Todo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "todo_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "GitHub",
        "url": "https://github.com/v420v/TodoApp"
    }
}